pointerevents/pointerevent_constructor.html

<!doctype html>
<html>
    <title>PointerEvent: Constructor test</title>
    <head>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <!-- Additional helper script for common checks across event types -->
        <script type="text/javascript" src="pointerevent_support.js"></script>	
    </head>
    <body onload="run()">
        <h1>PointerEvent: Dispatch custom event</h1>
        <h4>Test Description: This test checks if PointerEvent constructor works properly. This test is based on pointerover and pointerout events. Avoid real pointerover and pointerout for getting of the proper results.</h4>
        <script>
            var eventTested = false;
            var detected_pointertypes = {};
            
            setup({ explicit_done: true });
            add_completion_callback(showPointerTypes);
            
            function run() {
            	var target0 = document.getElementById("target0");
            	
            	var test_default_constructor = async_test("default constructor test");
            	var test_custom_constructor = async_test("custom constructor test");
            	
            	// set values for non-default constructor
            	var testBubbles = true;
            	var testCancelable = true;
            	var testPointerId = 42;
            	var testPointerType = 'pen';
            	var testClientX = 300;
            	var testClientY = 500;
            	var testWidth = 3;
            	var testHeight = 5;
            	var testTiltX = -45;
            	var testTiltY = 30;
            	var testPressure = 0.4;
            	var testIsPrimary = true;
            
            	var pointerEventCustom;
            	var pointerEventDefault;
            
            	on_event(target0, "pointerover", function(event) {
            		detected_pointertypes[ event.pointerType ] = true;
            		
            		if( (event.bubbles == testBubbles) &&
            			(event.cancelable == testCancelable) &&
            			(event.pointerId == testPointerId) &&
            			(event.pointerType == testPointerType) &&
            			(event.width == testWidth) &&
            			(event.height == testHeight) &&				
            			(event.clientX == testClientX) &&
            			(event.clientY == testClientY) &&
            			(event.tiltX == testTiltX) &&
            			(event.tiltY == testTiltY) &&
            			(event.pressure == testPressure) &&
            			(event.isPrimary == testIsPrimary)) {
            				test_custom_constructor.done();
            		}
            		else {
            			test_custom_constructor.step(function() {
            				assert_true(false,
            				"some of the values from constructor were not handled properly");
            			}, "some of the values from constructor were not handled properly");
            		}
            	});
            	
            	on_event(target0, "pointerout", function(event) {
            		if( (event.pointerId == 0) &&
            			(event.pointerType == "") &&
            			(event.width == 0) &&
            			(event.height == 0) &&				
            			(event.tiltX == 0) &&
            			(event.tiltY == 0) &&
            			(event.pressure == 0) &&
            			(event.isPrimary == false)) {
            				test_default_constructor.done();
            			}
            		else {
            			test_custom_constructor.step(function() {
            				assert_true(false,
            				"some of the default values were not handled properly");
            			}, "some of the default values were not handled properly");
            		}
            	});
            	
            	if(typeof(PointerEvent) != "undefined") {
            		pointerEventCustom = new PointerEvent("pointerover",
            		   {bubbles: testBubbles, 
            			cancelable: testCancelable, 
            			pointerId: testPointerId,
            			pointerType: testPointerType,
            			width: testWidth,
            			height: testHeight,
            			clientX: testClientX,
            			clientY: testClientY,
            			tiltX: testTiltX,
            			tiltY: testTiltY,
            			pressure: testPressure,
            			isPrimary: testIsPrimary
            			});
                    
                    // A PointerEvent created with a PointerEvent constructor must have all its attributes set to the corresponding values provided to the constructor.
                    // For attributes where values are not provided to the constructor, the corresponding default values must be used. 
                    // TA: 12.1
            		target0.dispatchEvent(pointerEventCustom);
            		
            		pointerEventDefault = new PointerEvent("pointerout");
            		
            		target0.dispatchEvent(pointerEventDefault);
            	}
            }
            
        </script>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
        </div>
        <div id="log"></div>
    </body>
</html>
